+++ core/src/main/java/com/ashelkov/owg/wallet/AlgorandWallet.java
@@ -9,33 +9,9 @@ import static com.ashelkov.owg.bip.Constants.BIP44_PURPOSE;
 
 public class AlgorandWallet extends SingleCoinWallet {
 
-    public static final Coin COIN = Coin.ALGO;
     public static final int PURPOSE = BIP44_PURPOSE;
 
     public AlgorandWallet(List<BIP44Address> derivedAddresses) {
-        super(derivedAddresses);
-    }
-
-    @Override
-    public String getIdentifier() {
-        return COIN.toString();
-    }
-
-    @Override
-    public String toString() {
-
-        StringBuilder result = new StringBuilder();
-
-        // append coin name
-        result.append(COIN);
-        result.append(':');
-
-        // append addresses
-        for (BIP44Address address : addresses) {
-            result.append('\n');
-            result.append(address.toString());
-        }
-
-        return result.toString();
+        super(derivedAddresses, Coin.ALGO);
     }
 }